+++ src/main/java/com/team03/ticketmon/payment/repository/PaymentRepository.java
@@ -3,9 +3,9 @@ package com.team03.ticketmon.payment.repository;
 import java.util.List;
 import java.util.Optional;
 
+import com.team03.ticketmon.booking.domain.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team03.ticketmon.concert.domain.Booking;
 import com.team03.ticketmon.payment.domain.entity.Payment;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
